# apps/api-gateway/doc/openapi.yml
...
openapi: 3.0.3
info:
  title: Movie Service API
  version: 1.0.0
  description: Quản lý thông tin website đặt vé xem phim
servers:
  - url: /api/{version}
    variables:
      version:
        default: v1
        enum:
          - v1
          - v2
...
paths:
...
  /cinema/{cinemaId}:
    patch:
      summary: Update cinema
      security:
        - bearerAuth: []
      tags:
        - Cinemas
      parameters:
        - name: cinemaId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateCinemaRequest'
      responses:
        '200':
          description: Cinema updated
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiSuccessResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/CinemaDetailResponse'
        '400':
          description: Lỗi request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiErrorResponse'

    delete:
      summary: Delete cinema
      security:
        - bearerAuth: []
      tags:
        - Cinemas
      parameters:
        - name: cinemaId
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Cinema deleted
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiSuccessResponse'
...
  /halls/cinema/{cinemaId}:
    get:
      summary: Get halls of a cinema
      tags:
        - Halls
      parameters:
        - name: cinemaId
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: List of halls
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiSuccessResponse'
                  - type: object
                    properties:
                      data:
                        type: array
                        items:
                          $ref: '#/components/schemas/HallSummaryResponse'
        '404':
          description: Cinema not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiErrorResponse'
  /halls/hall:
    post:
      summary: Create a hall
      tags:
        - Halls
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateHallRequest'
      responses:
        '201':
          description: Hall created
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiSuccessResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/HallDetailResponse'
        '400':
          description: Bad request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiErrorResponse'
  /halls/hall/{hallId}:
    get:
      summary: Get hall detail by ID
      tags:
        - Halls
      parameters:
        - name: hallId
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Hall detail
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiSuccessResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/HallDetailResponse'
        '404':
          description: Hall not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiErrorResponse'
    patch:
      summary: Update hall
      tags:
        - Halls
      parameters:
        - name: hallId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateHallRequest'
      responses:
        '200':
          description: Hall updated
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiSuccessResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/HallDetailResponse'
        '404':
          description: Hall not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiErrorResponse'
    delete:
      summary: Delete hall
      tags:
        - Halls
      parameters:
        - name: hallId
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Hall deleted
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiSuccessResponse'
        '404':
          description: Hall not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiErrorResponse'
  /halls/seat/{seatId}/status:
    patch:
      summary: Update seat status
      tags:
        - Halls
      parameters:
        - name: seatId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateSeatStatusRequest'
      responses:
        '200':
          description: Seat status updated
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiSuccessResponse'
                  - type: object
                    properties:
                      message:
                        type: string
                        example: Seat updated successfully
        '404':
          description: Seat not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiErrorResponse'

    # ----------------------
    # SHOWTIMES endpoints
    # ----------------------
...
  /promotions:
    get:
      summary: Get all promotions
      tags: [Promotions]
      parameters:
        - name: active
          in: query
          schema:
            type: string
        - name: type
          in: query
          schema:
            type: string
            enum: [PERCENTAGE, FIXED_AMOUNT, FREE_ITEM, POINTS]
      responses:
        '200':
          description: List of promotions
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiSuccessResponse'
                  - type: object
                    properties:
                      data:
                        type: array
                        items:
                          $ref: '#/components/schemas/PromotionResponse'
        '400':
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiErrorResponse'
    post:
      summary: Create promotion
      tags: [Promotions]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreatePromotionDto'
      responses:
        '201':
          description: Promotion created
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiSuccessResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/PromotionResponse'
        '400':
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiErrorResponse'
  /promotions/{id}:
    get:
      summary: Get promotion by ID
      tags: [Promotions]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Promotion details
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiSuccessResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/PromotionResponse'
        '404':
          description: Promotion not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiErrorResponse'
    put:
      summary: Update promotion
      tags: [Promotions]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdatePromotionDto'
      responses:
        '200':
          description: Promotion updated
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiSuccessResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/PromotionResponse'
        '404':
          description: Promotion not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiErrorResponse'
    delete:
      summary: Delete promotion
      tags: [Promotions]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Promotion deleted
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiSuccessResponse'
                  - type: object
                    properties:
                      message:
                        type: string
                        example: Promotion deleted successfully
        '404':
          description: Promotion not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiErrorResponse'
  /promotions/{id}/toggle-active:
    patch:
      summary: Toggle promotion active status
      tags: [Promotions]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Promotion status toggled
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiSuccessResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/PromotionResponse'
        '404':
          description: Promotion not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiErrorResponse'
  /promotions/code/{code}:
    get:
      summary: Get promotion by code
      tags: [Promotions]
      parameters:
        - name: code
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Promotion details
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiSuccessResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/PromotionResponse'
        '404':
          description: Promotion not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiErrorResponse'
  /promotions/validate/{code}:
    post:
      summary: Validate promotion code
      tags: [Promotions]
      parameters:
        - name: code
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ValidatePromotionDto'
      responses:
        '200':
          description: Validation result
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiSuccessResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/ValidatePromotionResponse'
        '400':
          description: Invalid details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiErrorResponse'
components:
  schemas:
...
    ErrorItem:
      type: object
      properties:
        code: { type: string }
        message: { type: string }
    ApiErrorResponse:
      type: object
      required: [success, message, errors, timestamp, path, statusCode]
      properties:
        success: { type: boolean, enum: [false] }
        message: { type: string }
        errors:
          type: array
          nullable: true
          items: { $ref: '#/components/schemas/ErrorItem' }
        timestamp: { type: string, format: date-time }
        path: { type: string }
        statusCode: { type: integer }
    ApiSuccessResponse:
      type: object
      properties:
        success:
          type: boolean
          example: true
        message:
          type: string
          nullable: true
        data:
          nullable: true
        meta:
          type: object
          nullable: true
        timestamp:
          type: string
          format: date-time
        path: { type: string }
      required: [success, message, timestamp]
...
    UpdateCinemaRequest:
      type: object
      description: Same fields as CreateCinemaRequest but optional
      properties:
        name:
          type: string
          nullable: true
        address:
          type: string
          nullable: true
        city:
          type: string
          nullable: true
        district:
          type: string
          nullable: true
        phone:
          type: string
          nullable: true
        email:
          type: string
          nullable: true
        website:
          type: string
          nullable: true
        latitude:
          type: number
          nullable: true
        longitude:
          type: number
          nullable: true
        description:
          type: string
          nullable: true
        amenities:
          type: array
          items:
            type: string
        facilities:
          type: object
          nullable: true
          additionalProperties: true
        images:
          type: array
          items:
            type: string
        virtualTour360Url:
          type: string
          nullable: true
        operatingHours:
          type: object
          nullable: true
          additionalProperties: true
        socialMedia:
          type: object
          nullable: true
          additionalProperties: true
        timezone:
          type: string
          nullable: true
...
    CinemaSummaryResponse:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        address:
          type: string
        city:
          type: string
        district:
          type: string
          nullable: true
        phone:
          type: string
          nullable: true
        email:
          type: string
          nullable: true
        rating:
          type: number
          nullable: true
        totalReviews:
          type: number
        images:
          type: array
          items:
            type: string
        status:
          type: string
          enum: [ACTIVE, MAINTENANCE, CLOSED]
    CinemaDetailResponse:
      type: object
      allOf:
        - $ref: '#/components/schemas/CinemaSummaryResponse'
        - type: object
          properties:
            website:
              type: string
              nullable: true
            latitude:
              type: number
              nullable: true
            longitude:
              type: number
              nullable: true
            description:
              type: string
              nullable: true
            amenities:
              type: array
              items:
                type: string
            facilities:
              type: object
              nullable: true
              additionalProperties: true
            virtualTour360Url:
              type: string
              nullable: true
            operatingHours:
              type: object
              nullable: true
              additionalProperties: true
            socialMedia:
              type: object
              nullable: true
              additionalProperties: true
            timezone:
              type: string
            createdAt:
              type: string
              format: date-time
            updatedAt:
              type: string
              format: date-time
...
    CreateHallRequest:
      type: object
      properties:
        cinemaId:
          type: string
          format: uuid
          example: '7b1db2ee-9a1f-4c22-a458-aa9cfae09e8b'
        name:
          type: string
          example: 'Hall 1'
        type:
          type: string
          enum: [STANDARD, PREMIUM, IMAX, FOUR_DX]
          example: STANDARD
        screenType:
          type: string
          nullable: true
          example: 'IMAX Laser'
        soundSystem:
          type: string
          nullable: true
          example: 'Dolby Atmos'
        features:
          type: array
          items:
            type: string
          example: ['3D', 'ATMOS']
        layoutType:
          type: string
          enum: [STANDARD, DUAL_AISLE, STADIUM]
          example: STANDARD
      required:
        - cinemaId
        - name
        - type
    UpdateHallRequest:
      type: object
      properties:
        name:
          type: string
          nullable: true
          example: 'Hall VIP 2'
        type:
          type: string
          enum: [STANDARD, PREMIUM, IMAX, FOUR_DX]
          nullable: true
          example: PREMIUM
        screenType:
          type: string
          nullable: true
        soundSystem:
          type: string
          nullable: true
        features:
          type: array
          nullable: true
          items:
            type: string
        # layoutType purposely omitted according to your DTO
    UpdateSeatStatusRequest:
      type: object
      properties:
        status:
          type: string
          enum: [ACTIVE, BROKEN, MAINTENANCE]
          example: BROKEN
      required:
        - status
    HallSummaryResponse:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
          example: 'Hall 1'
        type:
          type: string
          enum: [STANDARD, PREMIUM, IMAX, FOUR_DX]
        capacity:
          type: number
          example: 120
        rows:
          type: number
          example: 10
        status:
          type: string
          enum: [ACTIVE, MAINTENANCE, CLOSED]
          example: ACTIVE
        screenType:
          type: string
          nullable: true
        soundSystem:
          type: string
          nullable: true
    HallDetailResponse:
      allOf:
        - $ref: '#/components/schemas/HallSummaryResponse'
        - type: object
          properties:
            features:
              type: array
              items:
                type: string
              example: ['3D', 'ATMOS']
            cinemaId:
              type: string
              format: uuid
            layoutType:
              type: string
              enum: [STANDARD, DUAL_AISLE, STADIUM]
              example: STANDARD
            createdAt:
              type: string
              format: date-time
            updatedAt:
              type: string
              format: date-time

            seatMap:
              type: array
              items:
                type: object
                properties:
                  row:
                    type: string
                    example: 'A'
                  seats:
                    type: array
                    items:
                      type: object
                      properties:
                        id:
                          type: string
                          format: uuid
                        rowLetter:
                          type: string
                          example: 'A'
                        seatNumber:
                          type: number
                          example: 5
                        type:
                          type: string
                          enum: [STANDARD, VIP, COUPLE]
                          example: STANDARD
                        status:
                          type: string
                          enum: [ACTIVE, INACTIVE, MAINTENANCE]
                          example: ACTIVE

    # ----------------------
    # SHOWTIME SCHEMAS
    # ----------------------
...
    CreatePromotionDto:
      type: object
      required: [code, name, type, value, validFrom, validTo]
      properties:
        code: { type: string }
        name: { type: string }
        description: { type: string }
        type:
          type: string
          enum: [PERCENTAGE, FIXED_AMOUNT, FREE_ITEM, POINTS]
        value: { type: number }
        minPurchase: { type: number }
        maxDiscount: { type: number }
        validFrom: { type: string, format: date-time }
        validTo: { type: string, format: date-time }
        usageLimit: { type: number }
        usagePerUser: { type: number }
        applicableFor:
          type: array
          items: { type: string }
        conditions: { type: object }
        active: { type: boolean }
    UpdatePromotionDto:
      type: object
      properties:
        name: { type: string }
        description: { type: string }
        type:
          type: string
          enum: [PERCENTAGE, FIXED_AMOUNT, FREE_ITEM, POINTS]
        value: { type: number }
        minPurchase: { type: number }
        maxDiscount: { type: number }
        validFrom: { type: string, format: date-time }
        validTo: { type: string, format: date-time }
        usageLimit: { type: number }
        usagePerUser: { type: number }
        applicableFor:
          type: array
          items: { type: string }
        conditions: { type: object }
        active: { type: boolean }
    PromotionResponse:
      type: object
      properties:
        id: { type: string, format: uuid }
        code: { type: string }
        name: { type: string }
        description: { type: string }
        type:
          type: string
          enum: [PERCENTAGE, FIXED_AMOUNT, FREE_ITEM, POINTS]
        value: { type: number }
        minPurchase: { type: number }
        maxDiscount: { type: number }
        validFrom: { type: string, format: date-time }
        validTo: { type: string, format: date-time }
        usageLimit: { type: number }
        usagePerUser: { type: number }
        currentUsage: { type: number }
        applicableFor:
          type: array
          items: { type: string }
        conditions: { type: object }
        active: { type: boolean }
    ValidatePromotionItemDto:
      type: object
      required: [type, id, quantity]
      properties:
        type:
          type: string
          enum: [ticket, concession]
        id: { type: string }
        quantity: { type: number }
    ValidatePromotionDto:
      type: object
      required: [bookingAmount]
      properties:
        bookingAmount: { type: number }
        userId: { type: string }
        items:
          type: array
          items:
            $ref: '#/components/schemas/ValidatePromotionItemDto'
    ValidatePromotionResponse:
      type: object
      properties:
        valid: { type: boolean }
        promotion:
          $ref: '#/components/schemas/PromotionResponse'
        discountAmount: { type: number }
        finalAmount: { type: number }
        message: { type: string }
...
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT